# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LibraryOS API - Multi-tenant school bookstore management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (Super Admin only)
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant (Super Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant details
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/{tenantId}/toggle-status:
    post:
      operationId: toggleTenantStatus
      tags: [tenants]
      summary: Activate or deactivate a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========== PRODUCTS ==========
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        ownerName:
          type: string
        ownerEmail:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - ownerName
        - ownerEmail
        - isActive
        - createdAt
    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        address:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: string
        ownerPassword:
          type: string
      required:
        - name
        - slug
        - ownerName
        - ownerEmail
        - ownerPassword
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: string

    # PRODUCT
...